Makes test file for  Lesson component
